requirements.txt
--find-links https://prototyp3-dev.github.io/pip-wheels-riscv/wheels/

requests == 2.31.0
python-cartesi
